update auth api

+++ src/main/java/com/se/codingclub/controller/BrandController.java
@@ -18,11 +18,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.se.codingclub.dto.Auth;
 import com.se.codingclub.dto.BrandDTO;
 import com.se.codingclub.dto.ImageDTO;
 import com.se.codingclub.dto.ResponeMessage;
 import com.se.codingclub.entity.Brand;
 import com.se.codingclub.entity.Image;
+import com.se.codingclub.entity.User;
+import com.se.codingclub.service.AuthService;
 import com.se.codingclub.service.BrandService;
 import com.se.codingclub.service.ImageService;
 
@@ -35,7 +38,10 @@ public class BrandController {
 	private BrandService brandService;
 	@Autowired
 	private ImageService imageService;
-
+	@Autowired
+	private Auth tokenWarp;
+	@Autowired
+	private AuthService authService;
 	@GetMapping("/list")
 	public ResponseEntity<List<BrandDTO>> getBrands() {
 		Map<Brand, List<Image>> map = brandService.getListBrand();
@@ -93,14 +99,31 @@ public class BrandController {
 	}
 
 	@PostMapping("/new")
-	public Brand saveBrand(@RequestBody Brand brand) {
-
+	public Object saveBrand(@RequestBody Brand brand) {
+		String token  = tokenWarp.getToken();
+		if (token == null) {
+			return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED)
+					.body(new ResponeMessage("Please login to continue!"));
+		}
+		User user = authService.getUserByToken(token);
+		if(user.getRole().equals("admin") == false) {
+			return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponeMessage("Account does not have permission to perform this function!"));
+		}
 		return brandService.saveBrand(brand);
 
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<ResponeMessage> deleteBrand(@PathVariable String id) {
+		String token  = tokenWarp.getToken();
+		if (token == null) {
+			return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED)
+					.body(new ResponeMessage("Please login to continue!"));
+		}
+		User user = authService.getUserByToken(token);
+		if(user.getRole().equals("admin") == false) {
+			return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponeMessage("Account does not have permission to perform this function!"));
+		}
 		try {
 			brandService.deleteBrand(Integer.parseInt(id));
 			return ResponseEntity.ok().body(new ResponeMessage("Delete Success"));
